update homey.ink app js
major changes.
Most code comes from homeydash; with some changes.

I removed/disabled quick action function. No idea what it looks
like on a dash with, well, light switches or something like that
as I didn't test it - my favourites for the moment is only a
temperature sensor, which was identified to have a quick action
and with original code - it didn't show. Now it does. Touching
device entity changes value which is displayed, if device has
multiple properties.

This is experimental, I haven't tested this at all on flows for example;
But now it atleast shows devices, does not come with errors on js console
and has some *new* functionality.

This is a good point to continue developent. For someone else, unfortunately
I am busy with other apps (while writing this, Nilan heatpump app is in
early stage but in progress)

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>